requirements.txt
# requirements_compatible.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0